feat: Integrated Swagger + documentation outline
