+++ lib/db.js
@@ -1,5 +1,5 @@
 const sqlite3 = require("sqlite3");
-const { create_tables } = require("./setup-queries");
+const { create_tables } = require("./setupQueries");
 const parseCSV = require("../utils/parser");
 
 const db = new sqlite3.Database(":memory:", async (err) => {
@@ -10,9 +10,14 @@ const db = new sqlite3.Database(":memory:", async (err) => {
   console.log("Connected to SQLite3 database");
 
   setupDatabase();
+
   data = await parseCSV("./data/hdb-carpark-information-20220824010400.csv"); // replace this with the file parser of your choice
 
-  await runQuery(`INSERT INTO User VALUES (1, 'Test User')`); // Inserting test user for adding favorite
+  const userResult = await getQuery("SELECT * FROM User");
+  if(!userResult){
+    await runQuery(`INSERT INTO User VALUES (1, 'Test User')`); // Inserting test user for adding favorite
+  }
+
   await runQuery("BEGIN TRANSACTION");
   try {
     for (const row of data) {
@@ -24,8 +29,10 @@ const db = new sqlite3.Database(":memory:", async (err) => {
     console.error("Error inserting data:", err);
     await runQuery("ROLLBACK");
   }
+
 });
 
+// Function to parse free parking string into day, start_time and end_time fields
 function freeParkingParser(freeparking) {
   const freeParkingArray = [];
   try {
@@ -33,6 +40,7 @@ function freeParkingParser(freeparking) {
       // If the freeparking string contains "NO"
       return ["N.A.", "N.A.", "N.A."];
     }
+
     // If the freeparking string contains the day and time
     const temp1 = freeparking.split("FR");
     const day = temp1[0].replaceAll(" ", "");
@@ -88,6 +96,7 @@ function setupDatabase() {
   });
 }
 
+// Function to insert data into the normal form tables
 async function insertData(row) {
   let [
     car_park_no,
@@ -159,6 +168,7 @@ async function insertData(row) {
     );
     policyID = policyResult.lastID;
   }
+
   // For free parking table
   let freeparkingID;
   const freeParkingArray = freeParkingParser(free_parking);
@@ -193,6 +203,7 @@ async function insertData(row) {
   );
 }
 
+// Function to update data in the normal form tables
 async function updateData(row) {
   // Update the data in the database
   let [
@@ -307,8 +318,7 @@ async function updateData(row) {
   );
 }
 
-// Only deleting entries from the main table
-// Leaving the entries from the lookup tables as it is to avoid KeyNotFound errors
+// Function to delete data in the normal form tables
 async function deleteData(row) {
   // Delete the data in the database
   const car_park_no = row[0];
