+++ lib/setupQueries.js
