+++ package.json
@@ -15,6 +15,8 @@
     "express": "^4.21.2",
     "multer": "^1.4.5-lts.1",
     "nodemon": "^3.1.9",
-    "sqlite3": "^5.1.7"
+    "sqlite3": "^5.1.7",
+    "swagger-jsdoc": "^6.2.8",
+    "swagger-ui-express": "^5.0.1"
   }
 }
